requirements.txt
# Core (RAMT download script + project)
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.40

# ML / baselines
hmmlearn>=0.3.0
scikit-learn>=1.3.0
torch>=2.1.0
xgboost>=1.7.0
transformers>=4.40.0
peft>=0.10.0

# EDA / notebooks
matplotlib>=3.7.0
seaborn>=0.13.0
scipy>=1.10.0
statsmodels>=0.14.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Utils
tqdm>=4.66.0
PyYAML>=6.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Parquet IO
pyarrow>=15.0.0
